Break up edit person modal into smaller components
