+++ src/components/Contacts/ContactDetails/ContactDetailsTab/People/Items/EditPersonModal/ModalSectionContainer/ModalSectionContainer.tsx
@@ -1,0 +1,18 @@
+import React, { ReactNode } from 'react';
+import { Box, styled } from '@material-ui/core';
+
+interface ModalSectionContainerProps {
+  children: ReactNode;
+}
+
+const ModalSectionWrapper = styled(Box)(({ theme }) => ({
+  position: 'relative',
+  padding: theme.spacing(0, 6),
+  margin: theme.spacing(2, 0),
+}));
+
+export const ModalSectionContainer: React.FC<ModalSectionContainerProps> = ({
+  children,
+}) => {
+  return <ModalSectionWrapper>{children}</ModalSectionWrapper>;
+};
